refactor(ffi): change ffi code structure for bindgen
