+++ safe_api_ffi/Cargo.toml
@@ -41,10 +41,17 @@ self_update = "0.5.1"
 ffi_utils = "~0.12.0"
 maidsafe_utilities = "~0.18.0"
 
+[build-dependencies]
+safe_bindgen = { version = "~0.13.1", optional = true }
+rust_sodium = "~0.10.2"
+safe-nd = "~0.3.0"
+unwrap = "~1.2.1"
+
 [features]
 mock-network = ["safe_api/mock-network"]
 fake-auth = ["safe_api/fake-auth"]
 scl-mock = ["safe_api/scl-mock"]
+bindings = ["safe_bindgen"]
 
 [dev-dependencies]
 pretty_assertions = "0.6.1"
