+++ safe_api_ffi/build.rs
@@ -37,8 +37,8 @@ mod bindings {
     }
 
     fn gen_bindings_csharp() {
-        let target_dir = Path::new("bindings/csharp/safe_cli");
-
+        let target_dir = Path::new("../bindings/csharp/safe_api");
+        
         let mut bindgen = unwrap!(Bindgen::new());
         let mut lang = LangCSharp::new();
 
@@ -80,7 +80,7 @@ mod bindings {
         lang.reset_filter(FilterMode::Blacklist);
 
         let mut outputs = HashMap::new();
-        bindgen.source_file("src/lib.rs");
+        bindgen.source_file("lib.rs");
         bindgen.compile_or_panic(&mut lang, &mut outputs, true);
         apply_patches(&mut outputs);
         bindgen.write_outputs_or_panic(target_dir, &outputs);
