+++ safe_api_ffi/ffi/helpers.rs
