+++ safe_api_ffi/ffi/mod.rs
@@ -9,7 +9,7 @@ pub mod ffi_structs;
 
 use ffi_utils::{catch_unwind_cb, from_c_str, FfiResult, OpaqueCtx, FFI_RESULT_OK};
 use std::os::raw::{c_char, c_void};
-use crate::api::{Safe, ResultReturn};
+use safe_api::{Safe, ResultReturn};
 use helpers::{from_c_str_to_str_option};
 
 #[no_mangle]
