+++ safe_api_ffi/ffi/wallet.rs
